chore: Retire ancien dossier, renommage
