+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/GetResultsByEmail.tsx
