+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/HowToAct.tsx
