+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/Results.tsx
