+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/getResultsByEmail/Confirmation.tsx
