+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/howToAct/RecommendedAction.tsx
