+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/results/CategoriesChart.tsx
