+++ src/app/(layout-with-navigation)/(simulation)/fin/_components/results/categoriesAccordion/SubcategoriesList.tsx
